# api-spec.yml
swagger: "2.0"
info:
  title: Task API
  version: "1.0.0"

basePath: /api
schemes:
  - http

tags:
  - name: Tasks
    description: CRUD + CSV Import/Export

paths:
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks (search + pagination)
      produces:
        - application/json
      parameters:
        - in: query
          name: q
          type: string
          required: false
          description: Search query (matches task content)
        - in: query
          name: page
          type: integer
          required: false
          default: 1
          minimum: 1
        - in: query
          name: per_page
          type: integer
          required: false
          default: 5
          minimum: 1
          maximum: 100
        - in: query
          name: sort
          type: string
          required: false
          enum: [asc, desc]
          default: desc
          description: Sort by created time (asc = first added, desc = newest first)
      responses:
        200:
          description: Paginated list of tasks
          schema:
            $ref: "#/definitions/PaginatedTasks"

    post:
      tags: [Tasks]
      summary: Create a task
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateTaskRequest"
      responses:
        201:
          description: Task created
          schema:
            $ref: "#/definitions/Task"
        400:
          description: Invalid payload

  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by id
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: One task
          schema:
            $ref: "#/definitions/Task"
        404:
          description: Not found

    put:
      tags: [Tasks]
      summary: Update task by id
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateTaskRequest"
      responses:
        200:
          description: Updated task
          schema:
            $ref: "#/definitions/Task"
        400:
          description: Invalid payload
        404:
          description: Not found

    delete:
      tags: [Tasks]
      summary: Delete task by id
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        204:
          description: Deleted
        404:
          description: Not found

  /tasks/import-csv:
    post:
      tags: [Tasks]
      summary: Import tasks from CSV
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: formData
          name: file
          type: file
          required: true
          description: CSV file. Supports header content,completed OR single-column rows.
      responses:
        200:
          description: Import summary
          schema:
            $ref: "#/definitions/ImportSummary"
        400:
          description: Bad request
        500:
          description: DB insert failed

  /tasks/export-csv:
    get:
      tags: [Tasks]
      summary: Export tasks to CSV
      produces:
        - text/csv
      parameters:
        - in: query
          name: q
          type: string
          required: false
          description: Optional search filter before exporting
      responses:
        200:
          description: CSV file download
          schema:
            type: string

definitions:
  Task:
    type: object
    properties:
      id:
        type: integer
        example: 1
      content:
        type: string
        example: "Buy groceries"
      completed:
        type: integer
        example: 0
      created:
        type: string
        example: "2025-12-23T10:30:00"

  CreateTaskRequest:
    type: object
    required: [content]
    properties:
      content:
        type: string
        example: "Read system design notes"
      completed:
        type: integer
        default: 0
        example: 0

  UpdateTaskRequest:
    type: object
    properties:
      content:
        type: string
        example: "Updated task text"
      completed:
        type: integer
        example: 1

  PaginatedTasks:
    type: object
    properties:
      q:
        type: string
        example: "groceries"
      sort:
        type: string
        example: "desc"
      page:
        type: integer
        example: 1
      per_page:
        type: integer
        example: 5
      total:
        type: integer
        example: 100
      pages:
        type: integer
        example: 20
      has_next:
        type: boolean
        example: true
      has_prev:
        type: boolean
        example: false
      items:
        type: array
        items:
          $ref: "#/definitions/Task"

  ImportSummary:
    type: object
    properties:
      inserted:
        type: integer
        example: 10
      skipped:
        type: integer
        example: 2
      errors:
        type: integer
        example: 0